fix(scan): serialize scene handles as integers
Unity 6 projects scene handles as wrapper structs through HotRepl, which
violated the existing integer evidence contracts before collection could
finish. Cast every emitted handle at the runtime boundary and record the
first attributable 0.16.2 candidate run.

+++ packages/runtime/src/runtime.ts
@@ -354,7 +354,7 @@ export class Runtime {
           return new {
             researchCharacter = character.CharacterData.CharacterName,
             frame = UnityEngine.Time.frameCount,
-            scene = new { name = scene.name, path = scene.path, handle = scene.handle, isLoaded = scene.isLoaded },
+            scene = new { name = scene.name, path = scene.path, handle = (int)scene.handle, isLoaded = scene.isLoaded },
             gameSceneNativeId = nativeScene == null ? (int?)null : nativeScene.ID
           };
         });
